backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
spacy==3.7.4
sentence-transformers==2.5.1
PyPDF2==3.0.1
python-docx==1.1.0
